// src/aws/enums/aws-error-message.enum.ts
// Enums for Cognito Errors
export enum CognitoErrorMessage {
    COGNITO_CREATE_USER_FAILED = 'COGNITO_CREATE_USER_FAILED',
    COGNITO_ADD_USER_TO_GROUP_FAILED = 'COGNITO_ADD_USER_TO_GROUP_FAILED',
    COGNITO_INITIATE_AUTH_FAILED = 'COGNITO_INITIATE_AUTH_FAILED',
    COGNITO_CONFIRM_SIGNUP_FAILED = 'COGNITO_CONFIRM_SIGNUP_FAILED',
    COGNITO_CHANGE_PASSWORD_FAILED = 'COGNITO_CHANGE_PASSWORD_FAILED',
    COGNITO_NEW_PASSWORD_CHALLENGE_FAILED = 'COGNITO_NEW_PASSWORD_CHALLENGE_FAILED',
    COGNITO_CHANGE_INITIAL_PASSWORD_FAILED = 'COGNITO_CHANGE_INITIAL_PASSWORD_FAILED',
    COGNITO_GET_USER_DATA_FAILED = 'COGNITO_GET_USER_DATA',
    COGNITO_DELETE_USER_FAILED = 'COGNITO_DELETE_USER_FAILED',
}
  
// Enums for S3 Errors
export enum S3ErrorMessage {
    S3_UPLOAD_FAILED = 'S3_UPLOAD_FAILED',
    S3_DOWNLOAD_FAILED = 'S3_DOWNLOAD_FAILED',
    S3_DELETE_FAILED = 'S3_DELETE_FAILED',
    S3_LIST_OBJECTS_FAILED = 'S3_LIST_OBJECTS_FAILED',
}
  
// Enums for DynamoDB Errors
export enum DynamoDBErrorMessage {
    DYNAMO_GET_ITEM_FAILED = 'DYNAMO_GET_ITEM_FAILED',
    DYNAMO_PUT_ITEM_FAILED = 'DYNAMO_PUT_ITEM_FAILED',
    DYNAMO_UPDATE_ITEM_FAILED = 'DYNAMO_UPDATE_ITEM_FAILED',
    DYNAMO_DELETE_ITEM_FAILED = 'DYNAMO_DELETE_ITEM_FAILED',
}